move impl of debug.h to src
